chatbot-ifood/requirements.txt
langchain
langchain-community
langchain-groq
langchain-core
faiss-cpu
python-dotenv
fastapi
uvicorn
